backend/requirements.txt
# Step 1
# install through Dockerfile (docling)

# Step 2 (force CPU torch)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2
torchvision==0.17.2

numpy==1.26.4
opencv-python==4.8.1.78

# Step 3
sentence-transformers==2.7.0
transformers==4.57.6

# Step 4
qdrant-client==1.16.2
fastapi==0.129.0
uvicorn[standard]==0.41.0
pypdf==6.7.1

# Step 5
langchain-core==1.2.13
langchain-ollama==1.0.1
langgraph==1.0.8

# Step 6
ollama==0.6.1


faiss-cpu==1.13.2